Fix some compilation warnings:  * src/login.c: "dereferencing type-punned pointer will break    strict-aliasing rules", add a variable indirection: ptr_pam_user.  * lib/commonio.c: do not initialize the sb stat structure.  * lib/pwio.c, lib/shadowio.c, lib/sgroupio.c, lib/groupio.c:    initialize the security context if WITH_SELINUX.  * lib/nscd.c: The service argument is not const (used in the exec*    parameters). This matches with the prototype definition.  * src/groupmems.c: Avoid ++i when i is also used in the same line.  * src/newusers.c: i is positive every time it is compared. Add    cast to unsigned int.  * src/nologin.c: Use a main() prototype with no arguments.  * libmisc/getdate.y: Initialize the type and value fields of the    terminating entry for each TABLE.  * libmisc/tz.c: Use "TZ=CST6CDT" as the default timezone.

+++ src/newusers.c
@@ -116,11 +116,11 @@ static int add_group (const char *name, const char *gid, gid_t * ngid)
 	if (gid[0] == '\0') {
 		i = 100;
 		for (pw_rewind (); (pwd = pw_next ());) {
-			if (pwd->pw_uid >= i)
+			if (pwd->pw_uid >= (unsigned int)i)
 				i = pwd->pw_uid + 1;
 		}
 		for (gr_rewind (); (grp = gr_next ());) {
-			if (grp->gr_gid == i) {
+			if (grp->gr_gid == (unsigned int)i) {
 				i = -1;
 				break;
 			}
@@ -133,7 +133,7 @@ static int add_group (const char *name, const char *gid, gid_t * ngid)
 		 */
 		i = atoi (gid);
 		for (gr_rewind (); (grp = gr_next ());)
-			if (grp->gr_gid == i)
+			if (grp->gr_gid == (unsigned int)i)
 				goto add_member;
 	} else
 		/*
@@ -149,7 +149,7 @@ static int add_group (const char *name, const char *gid, gid_t * ngid)
 	 */
 	if (i == -1) {
 		for (i = 100, gr_rewind (); (grp = gr_next ());)
-			if (grp->gr_gid >= i)
+			if (grp->gr_gid >= (unsigned int)i)
 				i = grp->gr_gid + 1;
 	}
